requirements.txt
acres==0.3.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.2
distro==1.9.0
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.17.0
Flask==3.1.0
Flask-WTF==1.2.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==3.0.2
multidict==6.1.0
narwhals==1.30.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.3
openai==0.28.0
openpyxl==3.1.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
PyMuPDF==1.25.3
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.1
pyxnat==1.6.3
rdflib==6.3.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.43.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
